Added test-lookup-functions.ps1
This represents the basic premise for how I am going to do testing with PowerPing.
Basically we are going to load the PowerPing assembly into the powershell test script, then we will access the internal functions that we want to test.
We can check the return values (a few of the lookup functions have been converted to return strings instead of sending stuff to stout, which is probably for the good to start splitting up display and functionality anyway) and perform psuedo unit tests.

I know we could setup actual unit tests in C# but I want to avoid getting stuck into the minutia of writing tests in proper OO so writing the tests in powershell makes them easy to run, fairly easy to write and also easier to debug and not get caught up on