Complete refactoring of Ping.cs Changed names in PingAttributes and PingResults to make more sense and added better comments
I had been putting it off a while but Ping.cs needed to be cleaned up, it was just a mega function that did everything, it wasn't very approachable to anyone (even me) and everytime I tried to add something to it it made me throw up a little.

So I finally cleaned it all up; properly split it up into actually methods (hopefully didn't split things up too much), streamlined some functions, kept the actually flow pretty much the same (if it ain't broke don't fix it) but the whole thing has been properly cleaned and labelled and should be much nicer to extend and even use outside of PowerPing

Next I need to properly test each of the functions to make sure it still works and the I will sort out the display code because I believe there is a huge overlap in responsibilities and too much display logic is actually in Ping.cs

I feel that every year I try and rewrite this stuff but this time I do feel it will be different and this is genuinely for the better. I have to make the foundations strong before I add the features I want to in a proper way.